requirements.txt
amqp==5.2.0
billiard==4.2.0
blinker==1.8.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cramjam==2.8.3
fastparquet==2024.2.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
fsspec==2024.3.1
itsdangerous==2.2.0
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
numpy
packaging==24.0
pandas
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
WTForms==3.1.2
flower==2.0.1
humanize==4.9.0
prometheus_client==0.20.0
tornado==6.4
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozendict==2.4.4
html5lib==1.1
idna==3.7
lxml==5.2.2
multitasking==0.0.11
peewee==3.17.5
platformdirs==4.2.2
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
webencodings==0.5.1
yfinance==0.2.40
joblib==1.4.2
scikit-learn
scipy
threadpoolctl==3.5.0
copulas==0.11.0
plotly==5.22.0
tenacity==8.3.0
contourpy
cycler==0.12.1
fonttools==4.52.1
kiwisolver==1.4.5
matplotlib
patsy==0.5.6
pillow==10.3.0
pyparsing==3.1.2
statsmodels
Cython==0.29.37
hdbscan==0.8.37
llvmlite
numba
pynndescent==0.5.13
tqdm==4.66.4
umap-learn==0.5.6
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Flask-Migrate==4.0.7
Mako==1.3.5
alembic==1.13.2
psycopg2-binary==2.9.9
